Rename `country` to `countries`.
Consistent with other pluralised naming.
