+++ app/controllers/__init__.py
@@ -1,9 +1,9 @@
 from litestar import Router
 
 from app.domain.authors import Author
-from app.domain.country import Country
+from app.domain.countries import Country
 
-from . import authors, country
+from . import authors, countries
 
 __all__ = ["create_router"]
 
@@ -11,6 +11,6 @@ __all__ = ["create_router"]
 def create_router() -> Router:
     return Router(
         path="/v1",
-        route_handlers=[authors.AuthorController, country.CountryController],
+        route_handlers=[authors.AuthorController, countries.CountryController],
         signature_namespace={"Author": Author, "Country": Country},
     )
